+++ lib/countly-bulk-user.js
@@ -14,7 +14,7 @@
  * //adding requests by user
  * var user = server.add_user({device_id:"my_device_id"});
  */
-
+var {getProperties} = require("../lib/countly-common.js");
 /**
  * @lends module:lib/countly-bulk-user
  * Initialize CountlyBulkUser object
@@ -706,23 +706,7 @@ function CountlyBulkUser(conf) {
         return query;
     }
 
-    /**
-     *  Retrieve only specific properties from object
-     *  @param {Object} orig - object from which to get properties
-     *  @param {Array} props - list of properties to get
-     *  @returns {Object} Object with requested properties
-     */
-    function getProperties(orig, props) {
-        var ob = {};
-        var prop;
-        for (var i = 0; i < props.length; i++) {
-            prop = props[i];
-            if (typeof orig[prop] !== "undefined") {
-                ob[prop] = orig[prop];
-            }
-        }
-        return ob;
-    }
+
 }
 
 module.exports = CountlyBulkUser;
