+++ lib/countly-bulk.js
@@ -24,6 +24,7 @@ var fs = require("fs"),
     https = require("https"),
     cluster = require("cluster"),
     BulkUser = require("./countly-bulk-user");
+var {getTimestamp, getProperties, stripTrailingSlash} = require("../lib/countly-common.js");
 
 /**
  * @lends module:lib/countly-bulk
@@ -471,18 +472,6 @@ function CountlyBulk(conf) {
         return str.join("&");
     }
 
-    /**
-     *  Removing trailing slashes
-     *  @memberof Countly._internals
-     *  @param {String} str - string from which to remove traling slash
-     *  @returns {String} modified string
-     */
-    function stripTrailingSlash(str) {
-        if (str.substr(str.length - 1) === "/") {
-            return str.substr(0, str.length - 1);
-        }
-        return str;
-    }
 
     /**
     *  Making HTTP request
@@ -568,31 +557,6 @@ function CountlyBulk(conf) {
         }
     }
 
-    /**
-     *  Get current timestamp
-     *  @returns {number} unix timestamp in seconds
-     */
-    function getTimestamp() {
-        return Math.floor(new Date().getTime() / 1000);
-    }
-
-    /**
-     *  Retrieve only specific properties from object
-     *  @param {Object} orig - object from which to get properties
-     *  @param {Array} props - list of properties to get
-     *  @returns {Object} Object with requested properties
-     */
-    function getProperties(orig, props) {
-        var ob = {};
-        var prop;
-        for (var i = 0; i < props.length; i++) {
-            prop = props[i];
-            if (typeof orig[prop] !== "undefined") {
-                ob[prop] = orig[prop];
-            }
-        }
-        return ob;
-    }
 
     /**
      *  Handle messages from forked workers
