+++ lib/countly.js
@@ -27,6 +27,8 @@ var fs = require("fs"),
     https = require("https"),
     cluster = require("cluster"),
     Bulk = require("./countly-bulk");
+var {getTimestamp, getProperties, stripTrailingSlash, generateUUID} = require("../lib/countly-common.js");
+
 
 var Countly = {};
 
@@ -1195,19 +1197,6 @@ Countly.Bulk = Bulk;
         return storeGet("cly_id", null) || generateUUID();
     }
 
-    /**
-     *  Generate random UUID value
-     *  @returns {String} random UUID value
-     */
-    function generateUUID() {
-        var d = new Date().getTime();
-        var uuid = "xxxxxxxx-xxxx-4xxx-yxxx-xxxxxxxxxxxx".replace(/[xy]/g, function(c) {
-            var r = (d + Math.random() * 16) % 16 | 0;
-            d = Math.floor(d / 16);
-            return (c === "x" ? r : (r & 0x3 | 0x8)).toString(16);
-        });
-        return uuid;
-    }
 
     /**
     *  Get metrics of the browser or config object
@@ -1240,13 +1229,7 @@ Countly.Bulk = Bulk;
         }
     }
 
-    /**
-     *  Get current timestamp
-     *  @returns {number} unix timestamp in seconds
-     */
-    function getTimestamp() {
-        return Math.floor(new Date().getTime() / 1000);
-    }
+
 
     /**
      *  Get unique timestamp in miliseconds
@@ -1413,18 +1396,6 @@ Countly.Bulk = Bulk;
         return str.join("&");
     }
 
-    /**
-     *  Removing trailing slashes
-     *  @memberof Countly._internals
-     *  @param {String} str - string from which to remove traling slash
-     *  @returns {String} modified string
-     */
-    function stripTrailingSlash(str) {
-        if (str.substr(str.length - 1) === "/") {
-            return str.substr(0, str.length - 1);
-        }
-        return str;
-    }
 
     /**
      *  Parsing host and port information from url
@@ -1447,24 +1418,6 @@ Countly.Bulk = Bulk;
             serverOptions.port = Number(host.slice(lastPos + 1, host.length));
         }
         return serverOptions;
-    }
-
-    /**
-     *  Retrieve only specific properties from object
-     *  @param {Object} orig - object from which to get properties
-     *  @param {Array} props - list of properties to get
-     *  @returns {Object} Object with requested properties
-     */
-    function getProperties(orig, props) {
-        var ob = {};
-        var prop;
-        for (var i = 0; i < props.length; i++) {
-            prop = props[i];
-            if (typeof orig[prop] !== "undefined") {
-                ob[prop] = orig[prop];
-            }
-        }
-        return ob;
     }
 
     /**
